feat(miro-client): add refresh token stub
